+++ api/src/main/java/me/bymartrixx/playerevents/api/mixin/PlayerDeathMixin.java
@@ -1,8 +1,10 @@
 package me.bymartrixx.playerevents.api.mixin;
 
 import me.bymartrixx.playerevents.api.event.PlayerDeathCallback;
+import me.bymartrixx.playerevents.api.event.PlayerFirstDeathCallback;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.stat.Stats;
 import net.minecraft.util.ActionResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -13,9 +15,14 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class PlayerDeathMixin {
     @Inject(at = @At(value = "TAIL"), method = "onDeath", cancellable = true)
     private  void onPlayerDeath(DamageSource source, CallbackInfo info) {
-        PlayerDeathCallback.EVENT.invoker().kill((ServerPlayerEntity) (Object) this, source);
+        ServerPlayerEntity player = (ServerPlayerEntity) (Object) this;
+        if (player.getStatHandler().getStat(Stats.CUSTOM.getOrCreateStat(Stats.DEATHS)) < 1) {
+            PlayerFirstDeathCallback.EVENT.invoker().firstDeath(player, source);
+        }
 
-        ActionResult result1 = io.github.bymartrixx.playerevents.api.event.PlayerDeathCallback.EVENT.invoker().kill((ServerPlayerEntity) (Object) this, source);
+        PlayerDeathCallback.EVENT.invoker().kill(player, source);
+
+        ActionResult result1 = io.github.bymartrixx.playerevents.api.event.PlayerDeathCallback.EVENT.invoker().kill(player, source);
 
         if (result1 == ActionResult.FAIL) {
             info.cancel();